+++ prim_batch/__main__.py
@@ -66,8 +66,8 @@ class Logger(logging.Logger):
         if self.level == logging.NOTSET:
             self.setLevel(logging.WARNING if silent else logging.INFO)
 
-    def exception_or_error(self, e: Exception, args):
-        if not args or args.debug or args.test:
+    def exception_or_error(self, e: Exception):
+        if self.level == logging.NOTSET or self.level == logging.DEBUG:
             logger.exception(e)
         else:
             if hasattr(e, '__notes__'):
@@ -428,7 +428,7 @@ def main():
                 logger.error("Can't acquire lock on %s, probably already running", e.lock_file)
 
     except Exception as e:
-        logger.exception_or_error(e, args)
+        logger.exception_or_error(e)
 
     if print_stopped:
         logger.info("= STOPPED = %s %s", argv0, argvx)